requirements.txt
#For Automat-LLM
langchain_huggingface
langchain_community
langchain
transformers
sentence-transformers
pydantic
langchain_core
weaviate-client<5
faiss-cpu
numpy
groq
rich
langchain_groq
dia
playsound==1.2.2
#faiss-gpu

#For Automat-ImageGen
diffusers

#For automat-server
fastapi 
asyncio
uvicorn
uuid
